+++ Annex/src/utopia/annex/model/manifest/HasSchrodingerState.scala
@@ -2,7 +2,7 @@ package utopia.annex.model.manifest
 
 import utopia.annex.model.manifest.SchrodingerState.{Alive, Dead, Final, Flux}
 import utopia.flow.util.UncertainBoolean
-import utopia.flow.util.UncertainBoolean.CertainBoolean
+import utopia.flow.util.UncertainBoolean.{CertainlyFalse, CertainlyTrue}
 
 /**
   * A common trait for items which have a Schrödinger state, i.e. may be dead, alive or undetermined (Flux).
@@ -37,8 +37,8 @@ trait HasSchrodingerState
 	  *         Uncertain otherwise
 	  */
 	def isAlive: UncertainBoolean = state match {
-		case Alive => CertainBoolean(true)
-		case Dead => CertainBoolean(false)
+		case Alive => CertainlyTrue
+		case Dead => CertainlyFalse
 		case _ => UncertainBoolean
 	}
 	/**